Nav bar
I created a nav bar with the intention of adding more web pages. With the creation of the nav bar I added and completed the styles. I also went through and beautified the code to make it easier to read. Next I will start to do the html for css.